Added delete account confirmation modal.

+++ AccountManager.Blazor/Components/AccountListItem.razor.cs
@@ -11,10 +11,8 @@ namespace AccountManager.Blazor.Components
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public Account Account { get; set; }
 #pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
-
         [Parameter]
         public bool RenderButtons { get; set; }
-
         [Parameter, EditorRequired]
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public Action ReloadList { get; set; }
@@ -29,6 +27,7 @@ namespace AccountManager.Blazor.Components
         public IAccountService AccountService { get; set; }
 #pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
 
+        ConfirmationRequest? deleteAccountConfirmationRequest = null;
         bool loginDisabled = false;
         string loginBtnStyle => loginDisabled ? "color:darkgrey; pointer-events: none;" : "";
         LineChart<CoordinatePair>? lineChart;
@@ -165,8 +164,21 @@ namespace AccountManager.Blazor.Components
 
         public void Delete()
         {
-            AccountService.RemoveAccount(Account);
-            ReloadList();
+            deleteAccountConfirmationRequest = new ConfirmationRequest()
+            {
+                Callback = (userConfirmed) =>
+                {
+                    if (userConfirmed)
+                    {
+                        AccountService.RemoveAccount(Account);
+                        ReloadList();
+                    }
+
+                    deleteAccountConfirmationRequest = null;
+                    InvokeAsync(() => StateHasChanged());
+                },
+                RequestMessage = "Are you sure you want to delete this account? This can NOT be undone."
+            };
         }
     }
 }
